tools/taskCybozuTopmessage/functions/requirements.txt
Flask==1.0.2
requests
BeautifulSoup4
html5lib
sendgrid
google-cloud-datastore==1.12.0